% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aca.medicaid.expansion-package.R
\docType{package}
\name{aca.medicaid.expansion-package}
\alias{aca.medicaid.expansion}
\alias{aca.medicaid.expansion-package}
\title{aca.medicaid.expansion: An R data package containing the status of state action on the ACA 
  Medicaid expansion decision}
\description{
A dataset containing states' decisions about adopting the Medicaid 
  expansion under the ACA. From the Kaiser Family Foundation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jjchern/aca.medicaid.expansion}
  \item Report bugs at \url{https://github.com/jjchern/aca.medicaid.expansion/issues}
}

}
\author{
\strong{Maintainer}: JJ Chen \email{jiajia.chern@gmail.com} (\href{https://orcid.org/0000-0001-8482-8398}{ORCID})

}
\keyword{internal}
